[FIX]デプロイ用修正

+++ db/migrate/20200604021755_create_belongs.rb
@@ -1,8 +1,8 @@
 class CreateBelongs < ActiveRecord::Migration[5.2]
   def change
     create_table :belongs do |t|
-	  t.references :user, foreign_key: true, type: :integer
-	  t.references :team, foreign_key: true, type: :integer
+	  t.references :user, foreign_key: true, type: :bigint
+	  t.references :team, foreign_key: true, type: :bigint
 	  t.boolean :belong, default: false
       t.timestamps
     end
