+++ db/migrate/20200604022616_create_chats.rb
@@ -1,8 +1,8 @@
 class CreateChats < ActiveRecord::Migration[5.2]
   def change
     create_table :chats do |t|
-	  t.references :user, foreign_key: true, type: :integer
-	  t.references :room, foreign_key: true, type: :integer
+	  t.references :user, foreign_key: true, type: :bigint
+	  t.references :room, foreign_key: true, type: :bigint
 	  t.text :message
       t.timestamps
     end
